added jwt and trying to see if I can host

+++ resources/trainers/routes.py
@@ -1,8 +1,9 @@
 from flask import request
 from flask.views import MethodView
+from flask_jwt_extended import get_jwt_identity, jwt_required
 from flask_smorest import abort
 from sqlalchemy.exc import IntegrityError
-from schemas import TrainerSchema, TrainerSchemaNested, UpdateTrainerSchema, PokemonSchema, DeleteTrainerSchema
+from schemas import TrainerSchema, TrainerSchemaNested, UpdateTrainerSchema, PokemonSchema, AuthTrainerSchema
 from . import bp
 from .TrainerModel import TrainerModel
 from db import trainers, pokemon
@@ -27,13 +28,29 @@ class TrainerList(MethodView):
             abort(400, message='PC Name already taken')
 
     # delete trainer
-    @bp.arguments(DeleteTrainerSchema)
+    @jwt_required()
+    @bp.arguments(AuthTrainerSchema)
     def delete(self, trainer_data):
-        trainer = TrainerModel.query.filter_by(pc_user=trainer_data['pc_user']).first()
-        if trainer and trainer.check_pc_password(trainer_data['pc_password']):
+        trainer_id = get_jwt_identity()
+        trainer = TrainerModel.query.get(trainer_id)
+        if trainer and trainer.pc_user == trainer['pc_user'] and trainer.check_pc_password(trainer_data['pc_password']):
             trainer.delete()
             return {'message': f'{trainer_data["pc_user"]} deleted'}, 202
         abort(400, message='PC Name or PC Password Invalid')
+
+    # edit a trainer
+    @jwt_required()
+    @bp.arguments(UpdateTrainerSchema)
+    def put(self, trainer_data, trainer_id):
+        trainer_id = get_jwt_identity()
+        trainer = TrainerModel.query.gget(trainer_id)
+        if trainer and trainer.check_pc_password(trainer_data['pc_password']):
+            try:
+                trainer.from_dict(trainer_data)
+                trainer.save()
+                return trainer
+            except IntegrityError:
+                abort(400, message='PC User already taken')
 
 
 @bp.route('/trainers/<trainer_id>')
@@ -42,20 +59,14 @@ class Trainer(MethodView):
     # get a trainer
     @bp.response(200, TrainerSchemaNested)
     def get(self, trainer_id):
-        trainer = TrainerModel.query.get_or_404(trainer_id, description='User Not Found')
-        return trainer
-
-    # edit a trainer
-    @bp.arguments(UpdateTrainerSchema)
-    def put(self, trainer_data, trainer_id):
-        trainer = TrainerModel.query.get_or_404(trainer_id, description='Trainer Not Found')
-        if trainer and trainer.check_pc_password(trainer_data['pc_password']):
-            try:
-                trainer.from_dict(trainer_data)
-                trainer.save()
-                return trainer
-            except IntegrityError:
-                abort(400, message='PC User already taken')
+        user = None
+        if trainer_id.isdigit():
+            trainer = TrainerModel.query.get(trainer_id)
+        if not trainer:
+            trainer = TrainerModel.query.filter_by(pc_user=trainer_id).first()
+        if trainer:
+            return trainer
+        abort(400, message='Please enter valid PC User')
 
 @bp.get('/trainers/<trainer_id>/pokemon')
 @bp.response(200, PokemonSchema(many=True))
@@ -65,12 +76,14 @@ def get_trainer_pokemon(trainer_id):
     trainer_pokemon = [pokemon for pokemon in pokemon.values() if pokemon['trainer_id'] == trainer_id]
     return trainer_pokemon, 200
 
-@bp.route('/trainer/register/<registering_id>/<registered_id>')
+@bp.route('/trainer/register/<registered_id>')
 class RegisterTrainer(MethodView):
 
     # register trainer
+    @jwt_required
     @bp.response(200, TrainerSchema(many=True))
-    def post(self,registering_id,registered_id):
+    def post(self,registered_id):
+        registering_id = get_jwt_identity()
         trainer = TrainerModel.query.get(registering_id)
         trainer_to_register = TrainerModel.query.get(registered_id)
         if trainer and trainer_to_register:
@@ -79,8 +92,10 @@ class RegisterTrainer(MethodView):
         abort(400, message='Invalid Trainer Info')
 
     # unregister trainer
+    @jwt_required
     @bp.response(200, TrainerSchema(many=True))
-    def put(self,registering_id,registered_id):
+    def put(self,registered_id):
+        registering_id = get_jwt_identity()
         trainer = TrainerModel.query.get(registering_id)
         trainer_to_unregister = TrainerModel.query.get(registered_id)
         if trainer and trainer_to_unregister:
